Move: MemberTag 도메인 생성 및 파일 위치 변경 (#5)
