+++ src/main/java/com/sw19/sofa/domain/memberTag/repository/MemberTagRepository.java
@@ -1,4 +1,4 @@
-package com.sw19.sofa.domain.member.repository;
+package com.sw19.sofa.domain.memberTag.repository;
 
 import com.sw19.sofa.domain.member.entity.MemberTag;
 import org.springframework.data.jpa.repository.JpaRepository;
